fix(world-protocol): migrate movement opcodes to 3.4.3 (brick B4)
Migrate the movement opcode-name fallout from the B1 enum replacement
(legacy Opcodes:uint16 -> split OpcodeClient/OpcodeServer:uint32).

Per-site direction was resolved against TC wotlk_classic
(MovementHandler.cpp / MovementPackets.h / Unit.cpp):
- recv handler bindings  -> CMSG_MOVE_* (e.g. CMSG_MOVE_JUMP,
  CMSG_MOVE_FALL_LAND, CMSG_MOVE_FORCE_*_SPEED_CHANGE_ACK,
  CMSG_MOVE_FORCE_(UN)ROOT_ACK, CMSG_MOVE_SET_COLLISION_HEIGHT_ACK,
  CMSG_MOVE_CHANGE_TRANSPORT, CMSG_MOVE_TELEPORT_ACK).
- PC self packets        -> SMSG_MOVE_SET_* / SMSG_MOVE_(UN)ROOT /
  SMSG_MOVE_SET_HOVERING / SMSG_MOVE_SET_FEATHER_FALL / etc.
- spline broadcasts      -> SMSG_MOVE_SPLINE_*.
- observer rebroadcasts  -> SMSG_MOVE_UPDATE_* / SMSG_MOVE_SKIP_TIME /
  SMSG_ON_MONSTER_MOVE / SMSG_MOVE_UPDATE.

Verified the B2 SpeedOpcodePair remap against TC: PC=SMSG_MOVE_SET_*,
NPC=SMSG_MOVE_SPLINE_SET_*, ACK_RESPONSE=SMSG_MOVE_UPDATE_* — correct,
left unchanged.

Replaced the removed `Opcodes` type at recv sites with OpcodeClient;
widened Player::UpdateFallInformationIfNeed opcode param to uint32 so
the 0x38xxxx CMSG value is not truncated.

Redesigned 3.4.3 flows (force-ack rebroadcast, compound SMSG_MULTIPLE_MOVES,
monster-move-transport merge, teleport/heartbeat broadcast) are mapped to
the nearest directionally-correct 3.4.3 opcode and marked
`TODO(3.4.3 brick-B): ... verify in Phase-1d`.

Incremental checkpoint: `game` still does not fully link — the remaining
~237 errors are the non-movement clusters (Misc/Character/Query/Item/
Groups/Spells/Calendar/Warden) owned by B5. All movement-cluster opcode
errors are resolved; error total 381 -> 237.

codestyle-cpp: pass.

+++ src/server/game/Entities/Player/Player.h
@@ -2369,7 +2369,7 @@ public:
     /*********************************************************/
     /***                 VARIOUS SYSTEMS                   ***/
     /*********************************************************/
-    void UpdateFallInformationIfNeed(MovementInfo const& minfo, uint16 opcode);
+    void UpdateFallInformationIfNeed(MovementInfo const& minfo, uint32 opcode);
     SafeUnitPointer m_mover;
     WorldObject* m_seer;
     std::set<Unit*> m_isInSharedVisionOf;
